# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigiBoard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface ChangePasswordBody {
  currentPassword?: string;
  newPassword: string;
}

export interface AuthUser {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  role: string;
  mustChangePassword: boolean;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  role: string;
  mustChangePassword: boolean;
  createdAt: string;
}

export interface CreateUserBody {
  username: string;
  /** @nullable */
  email?: string | null;
  role: string;
}

export interface UpdateUserBody {
  username?: string;
  /** @nullable */
  email?: string | null;
  role?: string;
  mustChangePassword?: boolean;
}

export type PatientPassages = { [key: string]: string };

export interface Patient {
  id: number;
  clientNum: string;
  nom: string;
  prenom: string;
  /** @nullable */
  dob?: string | null;
  /** @nullable */
  adresse?: string | null;
  /** @nullable */
  tel?: string | null;
  /** @nullable */
  sexe?: string | null;
  /** @nullable */
  medecinFamille?: string | null;
  /** @nullable */
  patho?: string | null;
  /** @nullable */
  psy?: string | null;
  /** @nullable */
  responsable?: string | null;
  /** @nullable */
  casemanager2?: string | null;
  /** @nullable */
  demande?: string | null;
  /** @nullable */
  datePremierContact?: string | null;
  /** @nullable */
  dateEntree?: string | null;
  /** @nullable */
  dateSortie?: string | null;
  agressivite: number;
  /** @nullable */
  article?: string | null;
  /** @nullable */
  curatelle?: string | null;
  /** @nullable */
  remarques?: string | null;
  board: string;
  /** @nullable */
  phase?: string | null;
  /** @nullable */
  boardEntryDate?: string | null;
  passages?: PatientPassages;
  /** @nullable */
  recoveryObjectifs?: string | null;
  /** @nullable */
  recoveryEtape?: string | null;
  /** @nullable */
  recoveryAction?: string | null;
  /** @nullable */
  infosRecoltees?: string | null;
  /** @nullable */
  motifIrrecevable?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePatientBody {
  nom: string;
  prenom: string;
  /** @nullable */
  dob?: string | null;
  /** @nullable */
  adresse?: string | null;
  /** @nullable */
  tel?: string | null;
  /** @nullable */
  sexe?: string | null;
  /** @nullable */
  medecinFamille?: string | null;
  /** @nullable */
  patho?: string | null;
  /** @nullable */
  psy?: string | null;
  /** @nullable */
  responsable?: string | null;
  /** @nullable */
  casemanager2?: string | null;
  /** @nullable */
  demande?: string | null;
  /** @nullable */
  datePremierContact?: string | null;
  /** @nullable */
  dateEntree?: string | null;
  /** @nullable */
  dateSortie?: string | null;
  agressivite?: number;
  /** @nullable */
  article?: string | null;
  /** @nullable */
  curatelle?: string | null;
  /** @nullable */
  remarques?: string | null;
  board: string;
}

export interface UpdatePatientBody {
  nom?: string;
  prenom?: string;
  /** @nullable */
  dob?: string | null;
  /** @nullable */
  adresse?: string | null;
  /** @nullable */
  tel?: string | null;
  /** @nullable */
  sexe?: string | null;
  /** @nullable */
  medecinFamille?: string | null;
  /** @nullable */
  patho?: string | null;
  /** @nullable */
  psy?: string | null;
  /** @nullable */
  responsable?: string | null;
  /** @nullable */
  casemanager2?: string | null;
  /** @nullable */
  demande?: string | null;
  /** @nullable */
  datePremierContact?: string | null;
  /** @nullable */
  dateEntree?: string | null;
  /** @nullable */
  dateSortie?: string | null;
  agressivite?: number;
  /** @nullable */
  article?: string | null;
  /** @nullable */
  curatelle?: string | null;
  /** @nullable */
  remarques?: string | null;
  board?: string;
}

export interface MovePatientBoardBody {
  board: string;
  date: string;
}

export type UpdatePassagesBodyPassages = { [key: string]: string };

export interface UpdatePassagesBody {
  passages: UpdatePassagesBodyPassages;
}

export interface UpdatePhaseBody {
  /** @nullable */
  phase: string | null;
}

export interface UpdateRecoveryBody {
  /** @nullable */
  objectifs?: string | null;
  /** @nullable */
  etape?: string | null;
  /** @nullable */
  action?: string | null;
}

export interface UpdateInfosRecolteesBody {
  /** @nullable */
  infosRecoltees: string | null;
}

export interface UpdateMotifIrrecevableBody {
  /** @nullable */
  motifIrrecevable: string | null;
}

export interface MeetingNote {
  id: number;
  patientId: number;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  texte?: string | null;
  createdAt: string;
}

export interface CreateNoteBody {
  /** @nullable */
  date?: string | null;
  /** @nullable */
  texte?: string | null;
}

export interface UpdateNoteBody {
  /** @nullable */
  date?: string | null;
  /** @nullable */
  texte?: string | null;
}

export interface HistoryEntry {
  id: number;
  patientId: number;
  date: string;
  action: string;
  /** @nullable */
  boardTo?: string | null;
}

export interface UpdateHistoryEntryBody {
  date: string;
}

export type StatsBoardCounts = { [key: string]: number };

export type StatsSexeCounts = { [key: string]: number };

export type StatsAggCounts = { [key: string]: number };

export type StatsAvgDurations = { [key: string]: number | null };

export interface PathoCount {
  patho: string;
  count: number;
}

export interface Stats {
  total: number;
  active: number;
  byClosed: number;
  byIrrecevable: number;
  boardCounts: StatsBoardCounts;
  sexeCounts: StatsSexeCounts;
  pathoCounts: PathoCount[];
  aggCounts: StatsAggCounts;
  avgDurations: StatsAvgDurations;
}

export interface ActRegion {
  id: number;
  nom: string;
  createdAt: string;
}

export interface CreateActRegionBody {
  nom: string;
}

export interface UpdateActRegionBody {
  nom: string;
}

export interface ActNote {
  id: number;
  regionId: number;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  texte?: string | null;
  createdAt: string;
}

export interface PhaseOption {
  val: string;
  label: string;
}

export interface Settings {
  psychiatrists: string[];
  casemanagers: string[];
  articles: string[];
  curatelles: string[];
  phases: PhaseOption[];
}

export interface UpdateSettingBody {
  value: string;
}

export type ListPatientsParams = {
  board?: string;
  search?: string;
};