# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DigiBoard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── USERS ─────────────────────────────────────────────────────────────────
...
  /patients/{id}/notes:
    get:
      operationId: listPatientNotes
      tags: [notes]
      summary: List meeting notes for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetingNote"
    post:
      operationId: createPatientNote
      tags: [notes]
      summary: Create a meeting note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingNote"
  /patients/{id}/notes/{noteId}:
    patch:
      operationId: updatePatientNote
      tags: [notes]
      summary: Update a meeting note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingNote"
    delete:
      operationId: deletePatientNote
      tags: [notes]
      summary: Delete a meeting note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── BOARD HISTORY ─────────────────────────────────────────────────────────
  /patients/{id}/history:
    get:
      operationId: listPatientHistory
      tags: [history]
      summary: List board history for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"
  /patients/{id}/history/{historyId}:
    patch:
      operationId: updatePatientHistoryEntry
      tags: [history]
      summary: Update a history entry date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: historyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHistoryEntryBody"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"

  # ─── STATS ─────────────────────────────────────────────────────────────────
...
  /act/regions:
    get:
      operationId: listActRegions
      tags: [act]
      summary: List ACT regions
      responses:
        "200":
          description: Regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActRegion"
    post:
      operationId: createActRegion
      tags: [act]
      summary: Create an ACT region
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActRegionBody"
      responses:
        "201":
          description: Created region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActRegion"
  /act/regions/{id}:
    patch:
      operationId: updateActRegion
      tags: [act]
      summary: Update a region
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActRegionBody"
      responses:
        "200":
          description: Updated region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActRegion"
    delete:
      operationId: deleteActRegion
      tags: [act]
      summary: Delete a region
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /act/regions/{id}/notes:
    get:
      operationId: listActRegionNotes
      tags: [act]
      summary: List notes for a region
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActNote"
    post:
      operationId: createActRegionNote
      tags: [act]
      summary: Create a note for a region
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActNote"
  /act/regions/{id}/notes/{noteId}:
    patch:
      operationId: updateActRegionNote
      tags: [act]
      summary: Update a region note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActNote"
    delete:
      operationId: deleteActRegionNote
      tags: [act]
      summary: Delete a region note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SETTINGS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Auth ──
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [newPassword]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        mustChangePassword:
          type: boolean
      required: [id, username, role, mustChangePassword]

    # ── Users ──
...
    MeetingNote:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        date:
          type: ["string", "null"]
        texte:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, patientId, createdAt]
    CreateNoteBody:
      type: object
      properties:
        date:
          type: ["string", "null"]
        texte:
          type: ["string", "null"]
    UpdateNoteBody:
      type: object
      properties:
        date:
          type: ["string", "null"]
        texte:
          type: ["string", "null"]

    # ── History ──
    HistoryEntry:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        date:
          type: string
        action:
          type: string
        boardTo:
          type: ["string", "null"]
      required: [id, patientId, date, action]
    UpdateHistoryEntryBody:
      type: object
      properties:
        date:
          type: string
      required: [date]

    # ── Stats ──
...
    ActRegion:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        createdAt:
          type: string
      required: [id, nom, createdAt]
    CreateActRegionBody:
      type: object
      properties:
        nom:
          type: string
      required: [nom]
    UpdateActRegionBody:
      type: object
      properties:
        nom:
          type: string
      required: [nom]
    ActNote:
      type: object
      properties:
        id:
          type: integer
        regionId:
          type: integer
        date:
          type: ["string", "null"]
        texte:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, regionId, createdAt]

    # ── Settings ──
...