# lib/api-zod/src/generated/types/createNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigiBoard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNoteBody {
  /** @nullable */
  date?: string | null;
  /** @nullable */
  texte?: string | null;
}