# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigiBoard API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  patientId: number;
  date: string;
  action: string;
  /** @nullable */
  boardTo?: string | null;
}