# lib/api-zod/src/generated/types/meetingNote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DigiBoard API
 * OpenAPI spec version: 0.1.0
 */

export interface MeetingNote {
  id: number;
  patientId: number;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  texte?: string | null;
  createdAt: string;
}